+++ src/model/BeanBusi_fullcut.java
@@ -3,8 +3,8 @@ package model;
 public class BeanBusi_fullcut {
 	private String fullcut_id;
 	private String busi_id;
-	private int fullcut_fullvalue;
-	private int fullcut_cut;
+	private double fullcut_fullvalue;
+	private double fullcut_cut;
 	private int fullcut_if;
 	public String getFullcut_id() {
 		return fullcut_id;
@@ -18,16 +18,16 @@ public class BeanBusi_fullcut {
 	public void setBusi_id(String busi_id) {
 		this.busi_id = busi_id;
 	}
-	public int getFullcut_fullvalue() {
+	public double getFullcut_fullvalue() {
 		return fullcut_fullvalue;
 	}
-	public void setFullcut_fullvalue(int fullcut_fullvalue) {
+	public void setFullcut_fullvalue(double fullcut_fullvalue) {
 		this.fullcut_fullvalue = fullcut_fullvalue;
 	}
-	public int getFullcut_cut() {
+	public double getFullcut_cut() {
 		return fullcut_cut;
 	}
-	public void setFullcut_cut(int fullcut_cut) {
+	public void setFullcut_cut(double fullcut_cut) {
 		this.fullcut_cut = fullcut_cut;
 	}
 	public int getFullcut_if() {
